Fix crud operations & product entity properties

+++ src/test/java/com/tunnelnetwork/KpOnlineStore/RepositoryTests/CartRepositoryIntegrationTests.java
@@ -52,7 +52,7 @@ public class CartRepositoryIntegrationTests {
     userList.add(user.getEmail());
 
     Product product = new Product();
-    product.setPrice(200);
+    product.setPrice((double) 200);
     product.setProductName("Cream");
 
     productList.add(product);
