+++ src/test/java/com/tunnelnetwork/KpOnlineStore/RepositoryTests/ReceiptRepositoryIntegrationTests.java
@@ -54,7 +54,7 @@ public class ReceiptRepositoryIntegrationTests {
     userList.add(user.getEmail());
 
     Product product = new Product();
-    product.setPrice(200);
+    product.setPrice((double) 200);
     product.setProductName("Cream");
 
     productList.add(product);
